+++ java-Course/Springboot/springboot_learning/src/main/java/com/example/demo/service/AddressService.java
@@ -12,7 +12,7 @@ public class AddressService {
 	@Autowired
 	private AddressRepository addressRepository;
 
-	public Address getAddress(int id) {
+	public Address getAddress(Long id) {
 		return addressRepository.findById(id).orElse(null);
 	}
 }
